refact app and clean code
